update tests and ci action
